requirements.txt
APScheduler==3.6.3
blinker==1.7.0
cachetools==4.2.2
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-telegram-bot==13.15
pytz==2024.1
six==1.16.0
tornado==6.1
tzdata==2024.1
tzlocal==5.2
Werkzeug==3.0.1
